requirements.txt
discord.py==1.6.0
black==21.8b0
mypy==0.910
